% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_legis_replace.R
\name{speech_legis_replace}
\alias{speech_legis_replace}
\title{Rename legislators}
\usage{
speech_legis_replace(tidy_speech, old, new, id = NULL)
}
\arguments{
\item{tidy_speech}{tibble class \code{puy}.}

\item{old}{old name legislator.}

\item{new}{new name legislator.}

\item{id}{id 'floorspeech'.}
}
\value{
data.frame.
}
\description{
allows to adjust the name of the legislators prior to compiling the data.
}
\examples{
\dontrun{
speech_check(out2, "A")
out <- speech_legis_replace(out, old = "ABADALA",  new = "ABDALA")
}

}
