% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_uncompiler.R
\name{speech_uncompiler}
\alias{speech_uncompiler}
\title{Speech uncompiler}
\usage{
speech_uncompiler(tidy_speech)
}
\arguments{
\item{tidy_speech}{data.}
}
\value{
data.frame.
}
\description{
It allows to undo the compilation of a floor speech.
}
\examples{
\donttest{
# url <- "http://bit.ly/35AUVF4"
# out <- speech_build(file = url, compiler = TRUE)
# out2 <- speech_uncompiler(out)
}
}
