% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_url.R
\name{speech_url}
\alias{speech_url}
\title{url vectors}
\usage{
speech_url(chamber, from, to, days = NULL)
}
\arguments{
\item{chamber}{chamber:
\describe{
         \item{S}{Camara de Senadores}
         \item{D}{Camara de Representantes (Diputados)}
         \item{A}{Asamblea General}
         \item{C}{Comision Permanente}
        }}

\item{from}{character vector. Date in YYYY-MM-DD format}

\item{to}{character vector. Date in YYYY-MM-DD format}

\item{days}{character vector. Date in YYYY-MM-DD format.}
}
\value{
character vector
}
\description{
Allows to create a vector of url to download within a period within a legislature.
}
\examples{
# speech_url(chamber = "D",
#            from    = "2015-02-15",
#            to      = "2015-03-15")
#
# speech_url(chamber = "D",
#            from    = "2015-02-15",
#            to      = "2015-03-15")
#
# speech_url(chamber = "D",
#            days   = "2015-02-15")
#
# speech_url(chamber = "D",
#            days    = c("2002-06-12", "2004-04-14"))
#

}
