% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_view.R
\name{speech_view}
\alias{speech_view}
\title{View control speech}
\usage{
speech_view(tidy_speech, legis = character(), view = FALSE)
}
\arguments{
\item{tidy_speech}{data.frame class \code{puy}.}

\item{legis}{name of the legislator.}

\item{view}{logical. If \code{TRUE} \code{View} displays datasets containing
legislators' interventions (\code{legis}). By default is \code{FALSE}.}
}
\value{
data.frame.
}
\description{
Allows to see the legislators' names with problems prior to compiling the data.
}
\examples{
\donttest{
url <- "http://bit.ly/35AUVF4"
out <- speech_build(file = url)
speech_view(tidy_speech = out, legis = c("ABDALA", "LAZO"), view = FALSE)
}
}
