% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_legis_replace.R
\name{speech_legis_replace}
\alias{speech_legis_replace}
\title{Rename legislators}
\usage{
speech_legis_replace(tidy_speech, old, new, id = NULL)
}
\arguments{
\item{tidy_speech}{data.frame class \code{puy}.}

\item{old}{old legislator's name.}

\item{new}{new legislator's name.}

\item{id}{id 'floor speech'.}
}
\value{
data.frame.
}
\description{
allows to modify the legislators' name prior to compiling the data.
}
\examples{
\dontrun{
speech_check(out2, "A")
out <- speech_legis_replace(out, old = "ABADALA",  new = "ABDALA")
}

}
