% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_recompiler.R
\name{speech_recompiler}
\alias{speech_recompiler}
\title{Speech recompiler}
\usage{
speech_recompiler(
  tidy_speech,
  compiler_by = c("legislator", "party", "legislature", "chamber")
)
}
\arguments{
\item{tidy_speech}{tibble.}

\item{compiler_by}{character vector. Variables for which you want to recompile.}
}
\value{
tibble.
}
\description{
It allows to recompile an object of the \code{puy} class, the datasets
    speech or a data.frame built with \code{speech_build} to which the
    variable political party was added.
}
\details{
The default compilation is that of \ code {speech_build (dat, compiler = TRUE}).
    This function allows you to recompile the data by different levels of aggregation:
    chamber only, only legislature or other opacities.
}
\examples{
\dontrun{
dat <- data(speech)
speech_recompiler(dat)
speech_recompiler(dat)
}
}
