% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech_url.R
\name{speech_url}
\alias{speech_url}
\title{url vectors}
\usage{
speech_url(chamber, legislature, from, to)
}
\arguments{
\item{chamber}{chamber:
\describe{
         \item{S}{Camara de Senadores}
         \item{D}{Camara de Representantes (Diputados)}
         \item{A}{Asamblea General}
         \item{C}{Comision Permanente}
        }}

\item{legislature}{legislature number}

\item{from}{character vector. Date in DD-MM-YYY format}

\item{to}{character vector. Date in DD-MM-YYY format}
}
\value{
character vector
}
\description{
Allows to create a vector of url to download within a period within a legislature.
}
\examples{
# speech_url(chamber     = "D",
#            legislature = 48,
#            from        = "15-02-2015",
#            to          = "15-03-2015")
}
\author{
Elina Gomez \email{elina.gomez@cienciassociales.edu.uy}
}
